pyproject.toml
[project]
name = "multiagent-workflow"
version = "0.1.0"
description = "A LangGraph-based multi-agent system with supervisor pattern"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
dependencies = [
    "langgraph>=0.2.0",
    "langchain>=0.2.0",
    "langchain-openai>=0.1.0",
    "langchain-community>=0.2.0",
    "langchain-experimental>=0.0.60",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "tavily-python>=0.3.0",
    "langfuse>=2.0.0",
]
requires-python = ">=3.9"
readme = "README.md"
license = {text = "MIT"}

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]

[project.scripts]
multiagent = "main:main"
workflow-demo = "main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
include = [
    "agents.py",
    "workflow.py", 
    "main.py",
    "run.py",
    "test_import.py"
]

[dependency-groups]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]

[tool.black]
line-length = 88
target-version = ['py39']

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
